The actor system, that will read batches from the file and then will write out.
The system is composed by a Superviser, a Master, a Writer, and a pool of Workers.
The Superviser it is the entry point, it will start the Master and then watch for his termination.
The Master will generate batches of date in the limit of the number of active jobs, and then will submit thous batches to the pool of workers.
 The Worker will make the request the the real service, and then will submit a WriteJob to the Writer.
 The Writer should send to ouptut the processed batches. If the batches that he has are not ordered, they will be placed in a buffer, until some ordering can be done.